Ensure log file path is configured before logging errors
* Now checks if the log file path is defined before writing errors to
the log file
* Moves the log file outside of the Blueprint and into postmaster.utils
