+++ postmaster/apiv1/configs.py
@@ -9,10 +9,11 @@ from flask import request
 from flask_login import login_required, current_user
 from postmaster import db
 from postmaster.models import Configs
+from postmaster.utils import json_logger
 from ..decorators import json_wrap, paginate
 from ..errors import ValidationError, GenericError
 from . import apiv1
-from utils import json_logger, is_config_update_valid
+from utils import is_config_update_valid
 
 
 @apiv1.route("/configs", methods=["GET"])
