+++ postmaster/utils.py
@@ -7,10 +7,12 @@ Purpose: General helper utils
 import ldap
 from struct import unpack
 from re import search, sub, IGNORECASE
+from json import dumps
+from datetime import datetime
+from os import getcwd
 from wtforms.validators import StopValidation as WtfStopValidation
 from postmaster import db, models, bcrypt
 from postmaster.errors import ValidationError
-from apiv1.utils import json_logger
 
 
 def row2dict(row):
@@ -62,6 +64,49 @@ def getAlias(source):
     if queriedAlias:
         return row2dict(queriedAlias)
     return None
+
+
+def maildb_auditing_enabled():
+    """ Returns a bool based on if mail db auditing is enabled
+    """
+    auditing_setting = models.Configs.query.filter_by(
+        setting='Mail Database Auditing').first().value
+    return auditing_setting == 'True'
+
+
+def login_auditing_enabled():
+    """ Returns a bool based on if mail db auditing is enabled
+    """
+    auditing_setting = models.Configs.query.filter_by(
+        setting='Login Auditing').first().value
+    return auditing_setting == 'True'
+
+
+def json_logger(category, admin, message):
+    """
+    Takes a category (typically error or audit), a log message and the responsible
+    user. It then appends it with an ISO 8601 UTC timestamp to a JSON formatted log file
+    """
+    log_path = models.Configs.query.filter_by(setting='Log File').first().value
+    if log_path and ((category == 'error') or
+       (category == 'audit' and maildb_auditing_enabled()) or
+       (category == 'auth' and login_auditing_enabled())):
+        try:
+            with open(log_path, mode='a+') as log_file:
+                log_file.write("{}\n".format(dumps(
+                    {
+                        'category': category,
+                        'message': message,
+                        'admin': admin,
+                        'timestamp': datetime.utcnow().isoformat() + 'Z'
+                    },
+                    sort_keys=True)))
+                log_file.close()
+        except IOError:
+            raise ValidationError(
+                'The log could not be written to  "{0}". \
+                Verify that the path exists and is writeable.'.format(
+                    getcwd().replace('\\', '/') + '/' + log_path))
 
 
 def add_default_configuration_settings():
@@ -129,6 +174,7 @@ def add_default_configuration_settings():
         db.session.commit()
     except Exception:
         db.session.rollback()
+
 
 def add_ldap_user_to_db(username, display_name):
     """ Adds an LDAP user stub in the Admins table of the database for flask_login
